+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Version 1.4.4 (2025-05-04)
+- Updated pyz80.py to fix utf-8 BOM issue, add -B (boot) and -x (hex) options.
+- changed command line to use -B instead of -I for DOS image.
+
 ## Version 1.4.3 (2024-11-02)
 - Updated pyz80.py to fix Python 2.x issues on Linux and macOS.